# scim-server/src/main/resources/gluu-scim-openapi.yaml
openapi: 3.0.1
info:
  title: SCIM API
  description: |
    Gluu SCIM 2.0 server API. Developers can think of SCIM as a REST API with endpoints exposing CRUD functionality
    (create, update, retrieve and delete) for identity management resources such as users, groups, and fido devices.
  license:
    name: Apache 2
    url: https://github.com/GluuFederation/scim/blob/master/LICENSE
  version: 4.3.0
servers:
- url: https://your.gluu.server.com/identity/restv1/scim/v2
...
paths:
...
  /Schemas:
    get:
      tags:
      - discovery
      operationId: get-schemas
      description: "Endpoint used to retrieve information about schemas supported\
        \ (see section \n4 of RFC 7644)\n"
      responses:
        200:
          description: Successful operation
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/SchemaListResponse'
        500:
          description: There was an unexpected failure executing the operation
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /Schemas/{uri}:
    get:
      tags:
      - discovery
      operationId: get-schema-by-uri
      description: Retrieves information about a specific resource
      parameters:
      - name: uri
        in: path
        description: Schema URI of a particular resource type
        required: true
        schema:
          type: string
          enum:
          - urn:ietf:params:scim:schemas:core:2.0:User
          - urn:ietf:params:scim:schemas:extension:gluu:2.0:User
          - urn:ietf:params:scim:schemas:core:2.0:Group
          - urn:ietf:params:scim:schemas:core:2.0:Fido2Device
          - urn:ietf:params:scim:schemas:core:2.0:FidoDevice
      responses:
        200:
          description: Successful operation
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/SchemaResource'
        500:
          description: There was an unexpected failure executing the operation
          content:
            application/scim+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      required:
      - status
      type: object
      properties:
        schemas:
          type: array
          items:
            type: string
            example: urn:ietf:params:scim:api:messages:2.0:Error
        status:
          type: string
          description: HTTP status code as string
        scimType:
          type: string
          description: A detail error keyword. See table 9 of RFC 7644
        detail:
          type: string
          description: A detailed human-readable message of the error
      description: See section 3.12 of RFC 7644
...
    SchemaAttribute:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
        subAttributes:
          type: array
          items:
            $ref: '#/components/schemas/SchemaAttribute'
        multiValued:
          type: boolean
        description:
          type: string
        required:
          type: string
        canonicalValues:
          type: array
          items:
            type: string
        caseExact:
          type: boolean
        mutability:
          type: string
        returned:
          type: string
        uniqueness:
          type: string
        referenceTypes:
          type: array
          items:
            type: string
      description: A recursive definition. Note how property subAttributes reuses
        SchemaAttribute for its definition
    SchemaResource:
      type: object
      properties:
        schemas:
          type: array
          items:
            type: string
            example: urn:ietf:params:scim:schemas:core:2.0:Schema
        id:
          type: string
        meta:
          type: object
          properties:
            resourceType:
              type: string
            location:
              type: string
        name:
          type: string
        description:
          type: string
        attributes:
          type: array
          items:
            $ref: '#/components/schemas/SchemaAttribute'
    BasicListResponse:
      type: object
      properties:
        schemas:
          type: array
          items:
            type: string
            example: urn:ietf:params:scim:api:messages:2.0:ListResponse
        totalResults:
          type: integer
          description: Total number of results returned by the search. The value may be larger than the number of resources returned due to pagination
        startIndex:
          type: integer
          description: The 1-based index of the first result in the current set of search results
        itemsPerPage:
          type: integer
          description: The number of resources returned in a results page
...
    SchemaListResponse:
      allOf:
      - $ref: '#/components/schemas/BasicListResponse'
      - type: object
      - type: object
        properties:
          Resources:
            type: array
            items:
              $ref: '#/components/schemas/SchemaResource'
...
  securitySchemes:
    scim_oauth:
      type: oauth2
      description: |
        Endpoints protected by a bearer token passed in the Authorization header. 
        It is assumed gluu-scim is configured to operate in oauth protection mode. Other protection
        modes may be available, please check SCIM docs
      flows:
        clientCredentials:
          tokenUrl: 'https://your.gluu.server.com/oxauth/restv1/token'
          scopes:
            https://gluu.org/scim/users.read: Query user resources
            https://gluu.org/scim/users.write: Modify user resources
            https://gluu.org/scim/groups.read: Query group resources
            https://gluu.org/scim/groups.write: Modify group resources
            https://gluu.org/scim/fido.read: Query fido resources
            https://gluu.org/scim/fido.write: Modify fido resources
            https://gluu.org/scim/fido2.read: Query fido 2 resources
            https://gluu.org/scim/fido2.write: Modify fido 2 resources
            https://gluu.org/scim/all-resources.search: Access the root .search endpoint 
            https://gluu.org/scim/bulk: Send requests to the bulk endpoint